refactor online retrieval functions from feature store to utils
