+++ sdk/python/feast/feature_server.py
@@ -97,9 +97,9 @@ def get_app(
 
             full_feature_names = body.get("full_feature_names", False)
 
-            response_proto = store._get_online_features(
+            response_proto = store.get_online_features(
                 features=features,
-                entity_values=body["entities"],
+                entity_rows=body["entities"],
                 full_feature_names=full_feature_names,
             ).proto
